Move workflow stuff from pkg to workflow package
pkg is really just another name for utils and thats a bad pattern to have.
We should have small concise packages over large expansive ones.
